\name{bitest}
\docType{data}
\alias{bitest}
\title{A testing example for bivariate imputation}
\description{
\preformatted{
From: Maria Valaste [mailto:maria.valaste@helsinki.fi] 
Sent: 10 December 2012 18:45
To: Jing Hua Zhao
Subject: RE: correction on previous email: Help for using R package PAN

Dear Jing hua Zhao,

Thank You very much! This was just what I intended.
Of course you can include the example in the documentation. It's very good idea.


Best regards,
Maria Valaste


Lainaus "Jing Hua Zhao" <JingHua.Zhao@mrc-epid.cam.ac.uk>:

> Dear Maria,
>
> I just had time to try out your data/code; is this close to what you intend?
>
... 
>
> If you don't mind I could include your example in the documentation to 
> let others share your experiment.
>
> Best wishes,
>
>
> Jing Hua
>

From: Maria Valaste 
[mailto:maria.valaste@helsinki.fi] 
Sent: 08 December 2012 19:09
To: Jing Hua Zhao
Subject: Help for using R package PAN

Dear Jing hua Zhao,

I'm trying to use R package PAN. For one Y variable I have managed succesfully to 
impute using package PAN. When I try to impute two Y variables, I have encountered a 
problem that I can't solve. Could please help me? I'm not expert on R program and 
I'm really desperate.

I have R version 2.15.2 (Platform: x86_64-w64-mingw32/x64 (64-bit)). Below is code 
that I have tried to run for two Y variables. There also the error message (Error: 
subscript out of bounds). I also attach the test data below after the R code.


Best regards,
Maria Valaste
University of Helsinki, Finland
}
}
\usage{data(bitest)}
\format{A data frame}
\source{Maria Valaste <maria.valaste@helsinki.fi>}
  
\keyword{datasets}
